Add retry logic to AWS Security Group
Summary:
There are five steps that should happen when we create a security group:

1. Create the group
2. Revoke the default egress rule
3. Add egress rules
4. Add ingress rules
5. Set tags

A Terraform "create" action consists of all 5, and an "update" action consists of 3-5.

This patch makes two changes:

* Wrap steps 2, 3, and 4 in retry logic (1 and 5 are already wrapped)
* Refactor the "create" logic so we don't recheck for SG existence between steps 2 and 3

This revision is based on the branch in the following PR: https://github.com/hashicorp/terraform/pull/6775. That branch fixes the internet-gateway flakiness we've been seeing, and also adds retry logic to the "set tags" step for all AWS resources.

Test Plan: make test

Reviewers: hurshal, tyler, areece, carl

Reviewed By: carl

Subscribers: engineering-list

JIRA Issues: AP-499

Differential Revision: https://grizzly.memsql.com/D16013